Add ProxyStringLocalizer base class
